@@ -18,7 +18,8 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 2.4.1+2041
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.1 <3.0.0"
+  flutter: ">=3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
