@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 MaterialColor createMaterialColor(String hexColorCode) {
   Color color = Color(int.parse("0xFF$hexColorCode"));
   List strengths = <double>[.05];
-  Map swatch = <int, Color>{};
+  Map<int, Color> swatch = {};
   final int r = color.red, g = color.green, b = color.blue;
 
   for (int i = 1; i < 10; i++) {
